requirements.txt
python-dateutil==2.9.0
requests==2.32.3